+++ __tests__/helpers.ts
@@ -37,6 +37,8 @@ import {
   TransferType,
   type TransferStatus,
 } from '@dailydotdev/schema';
+import { createClient, type ClickHouseClient } from '@clickhouse/client';
+import * as clickhouseCommon from '../src/common/clickhouse';
 
 export class MockContext extends Context {
   mockSpan: MockProxy<opentelemetry.Span> & opentelemetry.Span;
@@ -510,5 +512,31 @@ export const createMockNjordErrorTransport = ({
         };
       },
     });
+  });
+};
+
+export const mockClickhouseClientOnce = () => {
+  const clientMock = createClient({});
+
+  jest
+    .spyOn(clickhouseCommon, 'getClickHouseClient')
+    .mockImplementationOnce(() => {
+      return clientMock;
+    });
+
+  return clientMock;
+};
+
+export const mockClickhouseQueryJSONOnce = <T>(
+  client: ClickHouseClient,
+  data: T,
+): jest.SpyInstance => {
+  return jest.spyOn(client, 'query').mockImplementationOnce(() => {
+    return {
+      json: async () => {
+        return data;
+      },
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    } as any;
   });
 };
